REST API/Challenge 5 JWT Authentication/requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
password_strength
